Add files via upload
I couldn't upload the DB file as it is, because GitHub treated it as an empty file and canceled the upload.

Just, unzip and place it in the correct location. Check unRAID support thread for more details.

You can also create a SQLite database file (empty) by your own if you prefer.